+++ test/actions/bulk/bulk_update_manual_test.exs
@@ -36,20 +36,28 @@ defmodule Ash.Test.Actions.BulkUpdateManualTest do
         |> case do
           {:ok, record} ->
             record =
-              Ash.Resource.put_metadata(
-                record,
+              record
+              |> Ash.Resource.put_metadata(
                 :bulk_update_index,
                 changeset.context.bulk_update.index
+              )
+              |> Ash.Resource.put_metadata(
+                :bulk_action_ref,
+                changeset.context.bulk_update.ref
               )
 
             [{:ok, record} | results]
 
           {:ok, record, notifications} ->
             record =
-              Ash.Resource.put_metadata(
-                record,
+              record
+              |> Ash.Resource.put_metadata(
                 :bulk_update_index,
                 changeset.context.bulk_update.index
+              )
+              |> Ash.Resource.put_metadata(
+                :bulk_action_ref,
+                changeset.context.bulk_update.ref
               )
 
             [{:ok, record, notifications} | results]
@@ -80,20 +88,28 @@ defmodule Ash.Test.Actions.BulkUpdateManualTest do
         |> case do
           {:ok, record} ->
             record =
-              Ash.Resource.put_metadata(
-                record,
+              record
+              |> Ash.Resource.put_metadata(
                 :bulk_update_index,
                 changeset.context.bulk_update.index
+              )
+              |> Ash.Resource.put_metadata(
+                :bulk_action_ref,
+                changeset.context.bulk_update.ref
               )
 
             [{:ok, record} | results]
 
           {:ok, record, notifications} ->
             record =
-              Ash.Resource.put_metadata(
-                record,
+              record
+              |> Ash.Resource.put_metadata(
                 :bulk_update_index,
                 changeset.context.bulk_update.index
+              )
+              |> Ash.Resource.put_metadata(
+                :bulk_action_ref,
+                changeset.context.bulk_update.ref
               )
 
             [{:ok, record, %{notifications: notifications}} | results]
@@ -124,20 +140,28 @@ defmodule Ash.Test.Actions.BulkUpdateManualTest do
         |> case do
           {:ok, record} ->
             record =
-              Ash.Resource.put_metadata(
-                record,
+              record
+              |> Ash.Resource.put_metadata(
                 :bulk_update_index,
                 changeset.context.bulk_update.index
+              )
+              |> Ash.Resource.put_metadata(
+                :bulk_action_ref,
+                changeset.context.bulk_update.ref
               )
 
             [{:ok, record} | results]
 
           {:ok, record, _notifications} ->
             record =
-              Ash.Resource.put_metadata(
-                record,
+              record
+              |> Ash.Resource.put_metadata(
                 :bulk_update_index,
                 changeset.context.bulk_update.index
+              )
+              |> Ash.Resource.put_metadata(
+                :bulk_action_ref,
+                changeset.context.bulk_update.ref
               )
 
             [{:ok, record} | results]